feat: add home screen icons for probability and option tools
