+++ apps/picsa-apps/extension-app/src/app/pages/home/home.page.scss
@@ -54,11 +54,11 @@ footer {
 }
 
 .nav-icon {
-  width: 60px;
-  height: 60px;
+  width: 70px;
+  height: 70px;
 }
 .nav-text {
-  height: 1em;
+  height: 2em;
 }
 .share-button {
   padding: 0;
